BLT/final-pjt-back/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
idna==3.4
joblib==1.4.2
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
pillow==10.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.5.0
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4